+++ androidApp/src/main/java/io/redlink/more/app/android/observations/AndroidObservationFactory.kt
@@ -14,11 +14,11 @@ import android.Manifest
 import android.content.Context
 import io.redlink.more.app.android.observations.GPS.GPSObservation
 import io.redlink.more.app.android.observations.GPS.GPSService
-import io.redlink.more.app.android.observations.HR.Polar360AccObservation
-import io.redlink.more.app.android.observations.HR.Polar360HrObservation
-import io.redlink.more.app.android.observations.HR.Polar360PpiObservation
-import io.redlink.more.app.android.observations.HR.Polar360TempObservation
-import io.redlink.more.app.android.observations.HR.PolarHeartRateObservation
+import io.redlink.more.app.android.observations.Polar.Polar360AccObservation
+import io.redlink.more.app.android.observations.Polar.Polar360HrObservation
+import io.redlink.more.app.android.observations.Polar.Polar360PpiObservation
+import io.redlink.more.app.android.observations.Polar.Polar360TempObservation
+import io.redlink.more.app.android.observations.Polar.PolarHeartRateObservation
 import io.redlink.more.app.android.observations.accelerometer.AccelerometerObservation
 import io.redlink.more.app.android.services.sensorsListener.BluetoothStateListener
 import io.redlink.more.app.android.services.sensorsListener.GPSStateListener
